#159: [Refactor] client componentsとserver componentsとの役割分別
